Moving definitions out of local_plugins/__init__.py
